// frontend/src/backend.d.ts
export interface SasaranRecord {
    id: bigint;
    status: Status;
    alamat: string;
    nama: string;
    nomorIdentitas: string;
    kategori: Kategori;
    catatan?: string;
}
export interface UserProfile {
    name: string;
    role: string;
    email?: string;
}
export enum DistribusiStatus {
    pending = "pending",
    terdistribusi = "terdistribusi",
    dalamProses = "dalamProses",
    tidakTerkirim = "tidakTerkirim"
}
export enum Kategori {
    balita = "balita",
    tidakDitentukan = "tidakDitentukan",
    ibuMenyusui = "ibuMenyusui",
    ibuHamil = "ibuHamil"
}
export enum Status {
    aktif = "aktif",
    nonAktif = "nonAktif"
}